+++ src/sphinx_codeautolink/__init__.py
@@ -8,7 +8,7 @@ from .extension import SphinxCodeAutoLink, backref, directive
 from .extension.block import clean_ipython, clean_pycon
 from .extension.translation import MESSAGE_CATALOG_NAME
 
-__version__ = "0.18.0"
+__version__ = "0.18.1"
 
 
 def setup(app: Sphinx):