+++ docs/src/release_notes.rst
@@ -8,8 +8,8 @@ These release notes are based on
 sphinx-codeautolink adheres to
 `Semantic Versioning <https://semver.org>`_.
 
-Unreleased
-----------
+0.18.1 (2026-04-30)
+-------------------
 - Recognize aliased ``TYPE_CHECKING`` imports such as
   ``from typing import TYPE_CHECKING as _TYPE_CHECKING`` (:issue:`199`)